feat(options): update tests
